fixed bug, hql returned by hibernate on the summary dto, not returning integer, returns long and double

+++ server/allTheMethods/src/main/java/com/example/allTheMethods/dto/response/SummaryResponseDto.java
@@ -1,7 +1,7 @@
 package com.example.allTheMethods.dto.response;
 
 public record SummaryResponseDto(
-        int countTotalProblems,
-        float userAcceptanceRate
+        Long countTotalProblems,
+        Double userAcceptanceRate
 ) {
 }
